refactor(setup): remove legacy welcome command
Removes the legacy terminal welcome command and the welcome-profile
persistence helper: pi-native first-run onboarding and the catalog-driven
setup source picker have replaced the post-setup role-selection flow. The
shared ASCII intro moves to WelcomeIntro.ts for the setup flows that still
print it, and updateSettingsDefaultProfile moves to setup/SetupStarterSource
alongside the other settings helpers. Adds first-run runtime onboarding
coverage and updates setup/source-layout tests for catalog-driven setup.

Rebuilt from ec8d8f6 (PR #107) against current main: the SetupCommand edits
are re-applied on top of the #127 decomposition, and the pi-extension restart
behavior changes from the original commit are intentionally dropped — that
code moved to code/pi-extension (#120) and the restart design overlaps with
PR #119's pending handoff work.
