+++ code/cli/tests/unit/source-layout.test.ts
@@ -18,7 +18,6 @@ import { createProfileCommands } from '../../src/cli/commands/profile/Command.js
 import { createRunCommand } from '../../src/cli/commands/RunCommand.js';
 import { createSetupCommand } from '../../src/cli/commands/SetupCommand.js';
 import { createSyncCommand } from '../../src/cli/commands/SyncCommand.js';
-import { createWelcomeCommand } from '../../src/cli/commands/WelcomeCommand.js';
 import { createEmptyProfile } from '../../src/profiles/Profile.js';
 import { createProfileLoadPlan } from '../../src/profiles/ProfileLoader.js';
 import { mergeProfileStack } from '../../src/profiles/ProfileMerger.js';
@@ -58,14 +57,13 @@ describe('source layout scaffolding', () => {
       'run',
       'setup',
       'sync',
-      'welcome',
       'profile',
       'profile list',
       'profile create',
       'profile lint',
     ]);
-    expect(program.commands.map((command) => command.name())).toEqual(['run', 'setup', 'sync', 'welcome', 'profile']);
-    expect(program.commands.at(4)?.commands.map((command) => command.name())).toEqual(['list', 'create', 'lint']);
+    expect(program.commands.map((command) => command.name())).toEqual(['run', 'setup', 'sync', 'profile']);
+    expect(program.commands.at(3)?.commands.map((command) => command.name())).toEqual(['list', 'create', 'lint']);
     expect(describeCommandObject(createRunCommand())).toEqual({
       name: 'run',
       description: 'Assemble a profile compositeProfile and launch the selected agent CLI.',
@@ -74,18 +72,12 @@ describe('source layout scaffolding', () => {
     const standaloneProgram = new Command();
     createSetupCommand().register(standaloneProgram);
     createSyncCommand().register(standaloneProgram);
-    createWelcomeCommand().register(standaloneProgram);
     for (const command of createProfileCommands()) {
       command.register(standaloneProgram);
     }
 
-    expect(standaloneProgram.commands.map((command) => command.name())).toEqual([
-      'setup',
-      'sync',
-      'welcome',
-      'profile',
-    ]);
-    expect(standaloneProgram.commands.at(3)?.commands.map((command) => command.name())).toEqual([
+    expect(standaloneProgram.commands.map((command) => command.name())).toEqual(['setup', 'sync', 'profile']);
+    expect(standaloneProgram.commands.at(2)?.commands.map((command) => command.name())).toEqual([
       'list',
       'create',
       'lint',
